Redo the lore around the Ramscoop Injector, and add missing strings.
I am unhappy with this lore for a couple of reasons--a presumably maintenance-intensive fuel collector does not make sense on a class that mostly stays in port, and in general reversing a device that concentrates diffuse matter will not produced focused output.

Still probably a placeholder--I would like something a bit more epic for the ship.